Gui: move class Camera to own source files
