+++ src/Gui/View3DInventor.cpp
@@ -57,6 +57,7 @@
 #include "View3DInventor.h"
 #include "View3DSettings.h"
 #include "Application.h"
+#include "Camera.h"
 #include "Document.h"
 #include "FileDialog.h"
 #include "MainWindow.h"
